Revert "Revert "Jvenstad/fix config model inconsitency""

+++ configserver/src/main/java/com/yahoo/vespa/config/server/tenant/TenantRequestHandler.java
@@ -6,6 +6,7 @@ import java.util.Collection;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Optional;
+import java.util.OptionalLong;
 import java.util.Set;
 
 import com.yahoo.component.Version;
@@ -16,6 +17,7 @@ import com.yahoo.vespa.config.protocol.ConfigResponse;
 import com.yahoo.vespa.config.server.NotFoundException;
 import com.yahoo.vespa.config.server.application.ApplicationMapper;
 import com.yahoo.vespa.config.server.application.ApplicationSet;
+import com.yahoo.vespa.config.server.application.TenantApplications;
 import com.yahoo.vespa.config.server.rpc.ConfigResponseFactory;
 import com.yahoo.vespa.config.server.host.HostRegistries;
 import com.yahoo.vespa.config.server.host.HostRegistry;
@@ -29,6 +31,8 @@ import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.TenantName;
 import com.yahoo.vespa.config.server.monitoring.MetricUpdater;
 import com.yahoo.vespa.config.server.monitoring.Metrics;
+import com.yahoo.vespa.curator.Curator;
+import com.yahoo.vespa.curator.Lock;
 
 /**
  * A per tenant request handler, for handling reload (activate application) and getConfig requests for
@@ -44,23 +48,25 @@ public class TenantRequestHandler implements RequestHandler, ReloadHandler, Host
     private final TenantName tenant;
     private final List<ReloadListener> reloadListeners;
     private final ConfigResponseFactory responseFactory;
-
     private final HostRegistry<ApplicationId> hostRegistry;
     private final ApplicationMapper applicationMapper = new ApplicationMapper();
     private final MetricUpdater tenantMetricUpdater;
     private final Clock clock = Clock.systemUTC();
+    private final TenantApplications applications;
 
     public TenantRequestHandler(Metrics metrics,
                                 TenantName tenant,
                                 List<ReloadListener> reloadListeners,
                                 ConfigResponseFactory responseFactory,
-                                HostRegistries hostRegistries) {
+                                HostRegistries hostRegistries,
+                                Curator curator) { // TODO jvenstad: Merge this class with TenantApplications, and straighten this out.
         this.metrics = metrics;
         this.tenant = tenant;
-        this.reloadListeners = reloadListeners;
+        this.reloadListeners = List.copyOf(reloadListeners);
         this.responseFactory = responseFactory;
-        tenantMetricUpdater = metrics.getOrCreateMetricUpdater(Metrics.createDimensions(tenant));
-        hostRegistry = hostRegistries.createApplicationHostRegistry(tenant);
+        this.tenantMetricUpdater = metrics.getOrCreateMetricUpdater(Metrics.createDimensions(tenant));
+        this.hostRegistry = hostRegistries.createApplicationHostRegistry(tenant);
+        this.applications = TenantApplications.create(curator, this, tenant);
     }
 
     /**
@@ -93,26 +99,40 @@ public class TenantRequestHandler implements RequestHandler, ReloadHandler, Host
      *
      * @param applicationSet the {@link ApplicationSet} to be reloaded
      */
+    @Override
     public void reloadConfig(ApplicationSet applicationSet) {
-        setLiveApp(applicationSet);
-        notifyReloadListeners(applicationSet);
+        ApplicationId id = applicationSet.getId();
+        try (Lock lock = applications.lock(id)) {
+            if ( ! applications.exists(id))
+                return; // Application was deleted before activation.
+            if (applicationSet.getApplicationGeneration() != applications.requireActiveSessionOf(id))
+                return; // Application activated a new session before we got here.
+
+            setLiveApp(applicationSet);
+            notifyReloadListeners(applicationSet);
+        }
     }
 
     @Override
     public void removeApplication(ApplicationId applicationId) {
-        if (applicationMapper.hasApplication(applicationId, clock.instant())) {
-            applicationMapper.remove(applicationId);
-            hostRegistry.removeHostsForKey(applicationId);
-            reloadListenersOnRemove(applicationId);
-            tenantMetricUpdater.setApplications(applicationMapper.numApplications());
-            metrics.removeMetricUpdater(Metrics.createDimensions(applicationId));
+        try (Lock lock = applications.lock(applicationId)) {
+            if (applications.exists(applicationId))
+                return; // Application was deployed again.
+
+            if (applicationMapper.hasApplication(applicationId, clock.instant())) {
+                applicationMapper.remove(applicationId);
+                hostRegistry.removeHostsForKey(applicationId);
+                reloadListenersOnRemove(applicationId);
+                tenantMetricUpdater.setApplications(applicationMapper.numApplications());
+                metrics.removeMetricUpdater(Metrics.createDimensions(applicationId));
+            }
         }
     }
 
     @Override
     public void removeApplicationsExcept(Set<ApplicationId> applications) {
         for (ApplicationId activeApplication : applicationMapper.listApplicationIds()) {
-            if (! applications.contains(activeApplication)) {
+            if ( ! applications.contains(activeApplication)) {
                 log.log(LogLevel.INFO, "Will remove deleted application " + activeApplication.toShortString());
                 removeApplication(activeApplication);
             }
@@ -236,5 +256,7 @@ public class TenantRequestHandler implements RequestHandler, ReloadHandler, Host
             reloadListener.verifyHostsAreAvailable(tenant, newHosts);
         }
     }
+
+    TenantApplications applications() { return applications; }
 
 }
